backend/requirements.txt
# GIS2BIM OpenAnalysis - Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP Client for WMS/WMTS
httpx>=0.26.0
aiohttp>=3.9.0

# Image Processing
Pillow>=10.2.0
reportlab>=4.0.0

# PDF Generation
weasyprint>=60.0
pypdf>=3.17.0

# Geospatial
pyproj>=3.6.0
owslib>=0.29.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Caching
aiocache>=0.12.0
